requirements.txt
alembic==1.0.9
Babel==2.6.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-translate==1.4.0
googleapis-common-protos==1.5.9
grpcio==1.20.0
guess-language-spirit==0.5.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.2
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
